change Worker on CoroutineWorker
